requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
future==1.0.0
iso8601==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pyserial==3.5
PyYAML==6.0.1
schedule==1.2.1
Werkzeug==3.0.1
